Tested some more concerning Y axis
Calibration of accel seems to be the problem. As noted in KrisWinner's git, most likely seems to be due to the temperature compensation bit which isn't really documented. Although I implemented the calibration as documented in the Ivensense AN, I will just resort to software calibration for the accelerometer, since without hardware calibration everything seems to be fine.